% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{agg}
\alias{agg}
\title{Aggregate data (age groups)}
\usage{
agg(population, inc)
}
\arguments{
\item{population}{population matrix, with consecutive 
single years as columns}

\item{inc}{incidence matrix, single years in columns}
}
\value{
list with adjusted population and incidence data
}
\description{
Aggregates yearly to 5yr age data (sum)
}
