% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zSpline.R
\name{createZSpline}
\alias{createZSpline}
\title{Create zSpline}
\usage{
createZSpline(val, a = NULL, b = NULL, nKnots = NULL)
}
\arguments{
\item{val}{Vector or values to fit a spline to}

\item{a}{Starting value, might be NULL (lowest value)}

\item{b}{Ending value, might be NULL (highest value)}

\item{nKnots}{number of knots}
}
\description{
Allows to incorporate splines into INLA 
based analyses
}
