% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate model projections.}
\usage{
evaluate(obj, dir = NULL, fnPre = "")
}
\arguments{
\item{obj}{incAnalysis object}

\item{dir}{directory to create plots in}

\item{fnPre}{prefix for plots}
}
\value{
updated incAnalysis object
}
\description{
Extracts and calculates precision data 
and evaluates bias and coverage (within / outside credibility 
or confidence bands). Two-sided 95% bands/intervals are used.
}
