% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwProjCalc.R
\name{forwProj}
\alias{forwProj}
\title{Perform forward projection}
\usage{
forwProj(cases, n, method, text, constRates, population,
  roundFun = identity)
}
\arguments{
\item{cases}{case counts (e.g. incidence) as stored in the 
casesWidePred slot of an incAnalysis object (period: rows; age: columns)}

\item{n}{Row index of cases to fill prediction years with}

\item{method}{which method to use, parameters used for DescTools::BinomCI,
defaults to "jeffreys": "wald", "wilson", "agresti-coull", "jeffreys", 
"modified wilson", "modified jeffreys", "clopper-pearson", "arcsine", 
"logit", "witting" or "pratt"}

\item{text}{Freely usable text variable}

\item{constRates}{calculate constant rates, use last observation of cases 
otherwise}

\item{population}{population counts in wide format (rows: years, cols: age),
same structure as for cases}

\item{roundFun}{function to apply to predictet cases (e.g. identity,
floor, ceiling)}
}
\description{
Forward projection of last observations
}
