% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamCalc.R
\name{gamCalc}
\alias{gamCalc}
\title{Perform GAM analysis}
\usage{
gamCalc(method = "inla", ...)
}
\arguments{
\item{method}{use INLA when set to "inla", classical
approach otherwise}

\item{...}{additional parameters}
}
\description{
Models data using GAMs,
independent of an incClass instancd
}
