% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmCalc.R
\name{glmCalc}
\alias{glmCalc}
\title{INLA based GLM calculation}
\usage{
glmCalc(method = "inla", ...)
}
\arguments{
\item{method}{inla for INLA or any other for classical approaches}

\item{...}{additional parameters}
}
\value{
list with results
}
\description{
Wrapper for inlaCalc() or mlCalc(),
}
