% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incClass.R
\docType{methods}
\name{initialize,incClass-method}
\alias{initialize,incClass-method}
\title{Constructor incClass}
\usage{
\S4method{initialize}{incClass}(.Object, dateCreated = base::POSIXct,
  population = data.frame, cases = data.frame, nPred = numeric,
  nFit = numeric, predLast = numeric, gridFactor = numeric,
  entity = character, ...)
}
\arguments{
\item{.Object}{class instance}

\item{dateCreated}{Date of creation}

\item{population}{data.frame containing count data with years in
rows (first row: earliest year) and numeric ages in cols (increasing from 
left to right)}

\item{cases}{data.frame containing count data, same design as the
population data}

\item{nPred}{Prediction period. Calculated backwards from the earliest valid
(in population and incidence available) year.}

\item{nFit}{Model fitting period. Counted backwards from the earliest 
year from which the prediction period starts. For forward projection of 
last observation, the year previous to the start of the prediction
interval is selected.}

\item{predLast}{Return only last observation year predictions and metrics 
when using the metrics() function.}

\item{gridFactor}{Ratio of age-width and yearly structure. If data is given 
in yearly fashion and 1-year age groups, the grid factor is 1. For yearly data 
and 5-year age groups, the grid factor is 5.}

\item{entity}{Free text variable.}

\item{...}{additional parameters}
}
\description{
Constructor for new analysis object
}
