% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlaCalc.R
\name{inlaCalc}
\alias{inlaCalc}
\title{INLA based model fitting}
\usage{
inlaCalc(data, frm = as.formula(Y ~ offset(log(N))),
  family = "poisson", control.family = list(link = "log"),
  control.predictor = list(link = 1, compute = T),
  control.compute = list(dic = T, cpo = T, waic = T), text = "",
  class = "INLA")
}
\arguments{
\item{data}{data in long format as stored in the dataLongPred 
slot of an incAnalysis object}

\item{frm}{Model formula.}

\item{family}{distribution}

\item{control.family}{inla parameter}

\item{control.predictor}{inla parameter}

\item{control.compute}{inla parameter}

\item{text}{Free text variable}

\item{class}{specify the type of caller, e.g. GLM für runGLM,
GAM for runGAM()}

\item{...}{additional parameters}
}
\value{
list with results
}
\description{
Fits models using the INLA framework
}
