% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Calculates metrics for incClass instances.}
\usage{
metrics(obj)
}
\arguments{
\item{obj}{incClass instance}
}
\value{
data.frame containing metrics
}
\description{
Calculates and aggregates bias, coverage
and precision. 
Negative bias hints towards overestimation of the 
projection approach (nPred > nObs).
}
