% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmCalc.R
\name{mlCalc}
\alias{mlCalc}
\title{ML based GLM fitting}
\usage{
mlCalc(data, frm = as.formula(Y ~ offset(log(N))), family = "poisson",
  text = "", type = "link")
}
\arguments{
\item{data}{data in long format as stored in the dataLongPred 
slot of an incAnalysis object}

\item{frm}{Model formula.}

\item{family}{Distribution}

\item{text}{Free text variable}

\item{type}{type parameter of the predict function}
}
\value{
list with results
}
\description{
Fits GLMs using an ML approach
}
