% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitHisto.R
\name{pitHist}
\alias{pitHist}
\title{Create PIT histograms}
\usage{
pitHist(obj, dir = NULL, fnPre = "")
}
\arguments{
\item{obj}{incAnalysis object}

\item{dir}{output directory for plots}

\item{fnPre}{Filename prefix}
}
\description{
Creates PIT histograms for INLA 
fitted models
}
