% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{reformat}
\alias{reformat}
\title{Reformats data}
\usage{
reformat(cases, population, pred = 2, fit = 15)
}
\arguments{
\item{cases}{data.frame with increasingly ordered years per 
row and age per column}

\item{population}{data.frame for population data, same structure as 
cases}

\item{pred}{number of years to predict}

\item{fit}{number of years to fit models with}
}
\value{
list with adjusted long and wide format data
}
\description{
Assures that case and population matrices match,
creates long format data from supplied wide format data,
sets values to predict to NA. Needed for the various data 
representations in the incAnalysis objects. Only data within 
the prediction and fitting periods are retained.
}
