% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bapcCalc.R
\name{runBAPC}
\alias{runBAPC}
\title{Run BAPC analysis}
\usage{
runBAPC(obj, text = "")
}
\arguments{
\item{obj}{incClass object}

\item{text}{free text variable}
}
\description{
Perform BAPC projection on inClass object
}
