% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamCalc.R
\name{runGAM}
\alias{runGAM}
\title{Run GAM analysis}
\usage{
runGAM(obj, frm = as.formula(Y ~ offset(log(N))), family = "poisson",
  text = "")
}
\arguments{
\item{frm}{Model formula. Population (offset) N, observed data Y,
age AGE and period (year) PERIOD}

\item{family}{exponential family to use (e.g. poisson)}

\item{text}{freely usable text variable}
}
\value{
Update incAnalysis object.
}
\description{
Models data using GAMs.
}
