% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlaCalc.R
\name{runInla}
\alias{runInla}
\title{Run INLA analysis}
\usage{
runInla(obj, frm = as.formula(Y ~ offset(log(N))), family = "poisson",
  control.family = list(link = "log"), control.predictor = list(link = 1,
  compute = T), control.compute = list(dic = T, cpo = T, waic = T),
  text = "")
}
\arguments{
\item{frm}{Model formula.}

\item{control.family}{inla parameter}

\item{control.predictor}{inla parameter}

\item{control.compute}{inla parameter}

\item{text}{Free text variable}

\item{data}{data in long format as stored in the dataLongPred 
slot of an incAnalysis object}
}
\value{
Updated incAnalysis object
}
\description{
Add INLA based model analysis to an incAnalysi sobject
}
