% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamCalc.R
\name{runMlGAM}
\alias{runMlGAM}
\title{Run GAM analysis}
\usage{
runMlGAM(obj, frm = as.formula(Y ~ offset(log(N))), family = "poisson",
  text = "")
}
\arguments{
\item{obj}{incClass object}

\item{frm}{model formula}

\item{family}{exponential family to use (e.g. poisson)}

\item{text}{freely usable text variable}
}
\value{
Update incAnalysis object.
}
\description{
Models data using GAMs.
}
