% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleAggMetr.R
\name{scaleAggMetr}
\alias{scaleAggMetr}
\title{Scale and aggregate metrics for score calculation}
\usage{
scaleAggMetr(bias, sd, cvg, wBIAS = 1, wSD = 1, wCVG = 1, plot = T)
}
\arguments{
\item{bias}{Bias variable}

\item{sd}{Precision variable (standard variation)}

\item{cvg}{Coverage}

\item{wBIAS}{weighting factor for bias}

\item{wSD}{weighting factor for standard deviation}

\item{wCVG}{weighting factor for coverage}

\item{plot}{plot data}
}
\value{
list with score values and matrices
}
\description{
Performs multiple adjustments of coverage, bias 
and precision values to obtain a score value.
}
