% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bapcCalc.R
\name{bapcCalc}
\alias{bapcCalc}
\title{BAPC calc}
\usage{
bapcCalc(population, cases, gf, predict = list(npredict = 10, retro = T),
  text = "")
}
\arguments{
\item{population}{Population data}

\item{cases}{Incidence counts}

\item{gf}{grid factor: age groups are m times wider than the period interval}

\item{predict}{precict parameter of BAPC}

\item{text}{Free text}
}
\description{
perform a standard BAPC analysis
}
