% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwProjCalc.R
\name{forwProj}
\alias{forwProj}
\title{Perform forward projection}
\usage{
forwProj(cases, n, ci, text)
}
\arguments{
\item{cases}{case counts (e.g. incidence) as stored in the 
casesWidePred slot of an incAnalysis object (period: rows; age: columns)}

\item{n}{Row index of cases to fill prediction years with}

\item{ci}{Confidence Interva, can be W for Wilson or PC for Pearson Clopper}

\item{text}{Freely usable text variable}
}
\description{
Forward projection of last observations
}
