% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamCalc.R
\name{gamMlCalc}
\alias{gamMlCalc}
\title{GAM based predictions}
\usage{
gamMlCalc(data, frm = as.formula(Y ~ offset(log(N))), text = "",
  family = "poisson", type = "link")
}
\arguments{
\item{data}{data used for model fitting, in o form as stored in 
the dataLongPred slot of an incAnalysis object}

\item{frm}{Formula used for mgcv::gam}

\item{text}{free-text variable}

\item{family}{distribution}

\item{type}{type parameter used for the predict function}
}
\description{
Use GAMs (mgcv) package for predictions.
}
