% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwProjCalc.R
\name{runFwProj}
\alias{runFwProj}
\title{Run forward projection}
\usage{
runFwProj(obj, ci = "W", text = "")
}
\arguments{
\item{ci}{Confidence Interval. Can be "W" for Wilson or 
PC for Pearson-Clopper.}

\item{text}{Freely usable additional textfield}
}
\value{
Updated incAnalysis object
}
\description{
Applies the forward projection of last observations
on the incAnalysis object
}
