% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamCalc.R
\name{runGAM}
\alias{runGAM}
\title{Run GAM analysis}
\usage{
runGAM(method = "inla", ...)
}
\arguments{
\item{method}{Can be inla for INLA fitted models,
any other value for classical approaches}

\item{...}{additional parameters}
}
\description{
Run GAM analysis on incClass instance,
wrapper for runInla() or runMlGAM()

Models data using GAMs.
}
