% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamCalc.R
\name{runMlGAM}
\alias{runMlGAM}
\title{Run ML GAM analysis}
\usage{
runMlGAM(obj, frm = as.formula(Y ~ offset(log(N))), family = "poisson",
  text = "")
}
\arguments{
\item{obj}{incClass object}

\item{frm}{model formula}

\item{family}{distribution}

\item{text}{freely usable text variable}
}
\value{
Updated incClass object.
}
\description{
Models data using GAMs, ML approach.
}
